requirements.txt
langchain 
llama_index
openai
weaviate-client
python-dotenv
fire
fastapi
uvicorn
transformers
torch
runhouse
sentence_transformers